+++ src/Platform/Console/Commands/MakeChart.php
@@ -1,0 +1,51 @@
+<?php
+
+namespace Orchid\Platform\Console\Commands;
+
+use Illuminate\Console\GeneratorCommand;
+
+class MakeChart extends GeneratorCommand
+{
+    /**
+     * The console command name.
+     *
+     * @var string
+     */
+    protected $name = 'make:chart';
+
+    /**
+     * The console command description.
+     *
+     * @var string
+     */
+    protected $description = 'Create a new chart layout class';
+
+    /**
+     * The type of class being generated.
+     *
+     * @var string
+     */
+    protected $type = 'Chart';
+
+    /**
+     * Get the stub file for the generator.
+     *
+     * @return string
+     */
+    protected function getStub() : string
+    {
+        return DASHBOARD_PATH.'/resources/stubs/chart.stub';
+    }
+
+    /**
+     * Get the default namespace for the class.
+     *
+     * @param string $rootNamespace
+     *
+     * @return string
+     */
+    protected function getDefaultNamespace($rootNamespace) : string
+    {
+        return $rootNamespace.'\Layouts';
+    }
+}